Fix coverage of Mutant::Context namespace
* Fix Mutant::Context::Scope as only subclass of abstract
  Mutant::Context via onlining Scope semantics into Context.
