+++ lib/mutant.rb
@@ -131,7 +131,6 @@ require 'mutant/mutator/node/rescue'
 require 'mutant/mutator/node/match_current_line'
 require 'mutant/loader'
 require 'mutant/context'
-require 'mutant/context/scope'
 require 'mutant/scope'
 require 'mutant/subject'
 require 'mutant/subject/method'
